+++ src/main/java/me/efekos/simpler/items/FieldType.java
@@ -20,19 +20,14 @@
  * SOFTWARE.
  */
 
-package me.efekos.simpler.annotations;
+package me.efekos.simpler.items;
 
-import me.efekos.simpler.items.custom.CustomItem;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Used to separate the {@link org.bukkit.event.player.PlayerInteractEvent}s on a {@link CustomItem}.
- */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface RightClick {
+public enum FieldType {
+    STRING,
+    INTEGER,
+    DOUBLE,
+    LONG,
+    FLOAT,
+    BOOLEAN,
+    BYTE
 }